feat: add feature for distinguishing higher level components in ExpressionTranslator

+++ src/test/java/bytecode_generator/BytecodeGeneratorTest.java
@@ -38,7 +38,7 @@ class BytecodeGeneratorTest {
 
 		BytecodeGenerator bytecodeGenerator = new BytecodeGenerator();
 
-		int[] bytecode = bytecodeGenerator.generateBytecode(convertedTokens);
+		byte[] bytecode = bytecodeGenerator.generateBytecode(convertedTokens);
 
 		for (int i = 0; i < expectedOutput.length; i++) {
